backend/requirements.txt
# Optional extras required at runtime
email-validator

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
apscheduler==3.10.4
africastalking==1.2.6
twilio==8.10.3
pyserial==3.5
httpx==0.25.1
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
croniter==2.0.1

